........
  r3018 | danydb | 2010-04-16 12:45:48 +0200 (Fri, 16 Apr 2010) | 4 lines
  
  Quick_writing
  =============
  Advanced->direct writing remove the flag direct_write
........
  r3019 | danydb | 2010-04-16 13:11:06 +0200 (Fri, 16 Apr 2010) | 7 lines
  
  Advanced->Periode
  ==================
  * give the number of operation
  * do not propose to remove a used periode
  * fix cosmetic
  * add confirm for "cloture" and "remove"
........
  r3020 | danydb | 2010-04-16 15:31:03 +0200 (Fri, 16 Apr 2010) | 2 lines
  
  IPoste : add documentation : sample of code
........
  r3021 | danydb | 2010-04-16 15:31:32 +0200 (Fri, 16 Apr 2010) | 2 lines
  
  Iposte update doc : typo
........
  r3022 | danydb | 2010-04-16 21:12:47 +0200 (Fri, 16 Apr 2010) | 2 lines
  
  Add info for error
........
  r3023 | danydb | 2010-04-17 00:37:08 +0200 (Sat, 17 Apr 2010) | 2 lines
  
  ac_common.php add a function h2info
........
  r3024 | danydb | 2010-04-17 01:19:21 +0200 (Sat, 17 Apr 2010) | 2 lines
  
  Button is not printed
........
  r3025 | danydb | 2010-04-17 01:20:14 +0200 (Sat, 17 Apr 2010) | 2 lines
  
  extension :Logout button and name plugin is not printed
........
  r3026 | danydb | 2010-04-17 01:40:01 +0200 (Sat, 17 Apr 2010) | 1 line
  
  Appearance : improve disable html elt appearance
........
  r3027 | danydb | 2010-04-17 01:43:05 +0200 (Sat, 17 Apr 2010) | 1 line
  
  Periode : improve usuability
........
  r3028 | danydb | 2010-04-17 01:43:21 +0200 (Sat, 17 Apr 2010) | 1 line
  
  Fix several bugs in itva_popup
........
  r3029 | danydb | 2010-04-17 14:12:47 +0200 (Sat, 17 Apr 2010) | 2 lines
  
  New Manuel for windows
........
  r3030 | danydb | 2010-04-17 15:18:25 +0200 (Sat, 17 Apr 2010) | 6 lines
  
  
  Document modele
  ===============
  New : can decide if a document must be used with the ledger of sales, purchase or follow up
........
  r3031 | danydb | 2010-04-18 00:05:25 +0200 (Sun, 18 Apr 2010) | 1 line
  
  Add function tr
........
  r3032 | danydb | 2010-04-18 00:06:15 +0200 (Sun, 18 Apr 2010) | 1 line
  
  Correct style : you can now click on the selected menu
........
  r3033 | danydb | 2010-04-18 19:58:24 +0200 (Sun, 18 Apr 2010) | 3 lines
  
  
  scripts.js : add function deleteRow(tablename, object) to use with IButton
........

+++ include/class_periode.php
@@ -157,8 +157,10 @@ class Periode {
     $str_dossier=dossier::get();
 
     if ( $this->jrn_def_id==0 ) {
-      $Res=$this->cn->exec_sql("select p_id,to_char(p_start,'DD.MM.YYYY') as date_start,to_char(p_end,'DD.MM.YYYY') as date_end,p_central,p_closed,p_exercice
-  from parm_periode order by p_start,p_end");
+      $Res=$this->cn->exec_sql("select p_id,to_char(p_start,'DD.MM.YYYY') as date_start,to_char(p_end,'DD.MM.YYYY') as date_end,p_central,p_closed,p_exercice,
+(select count(jr_id) as count_op from jrn where jr_tech_per = p_id) as count_op
+  from parm_periode 
+order by p_start,p_end");
       $Max=Database::num_row($Res);
       echo '<TABLE ALIGN="CENTER">';
       echo "</TR>";
@@ -180,16 +182,22 @@ class Periode {
 	  $remove='<TD></TD>';
 	} else {
 	  $closed='<TD class="mtitle">';
-	  $closed.='<A class="mtitle" HREF="?p_action=periode&action=closed&p_per='.$l_line['p_id'].'&'.$str_dossier.'"> Cloturer</A>';
+	  $closed.='<A class="mtitle" HREF="?p_action=periode&action=closed&p_per='.$l_line['p_id'].'&'.$str_dossier.'" onclick="return confirm(\''._('Confirmez cloture').' ?\')"> Cloturer</A>';
 	  $change='<TD class="mtitle">';
 	  $change.='<A class="mtitle" HREF="?p_action=periode&action=change_per&p_per='.
 	    $l_line['p_id']."&p_date_start=".$l_line['date_start'].
 	    "&p_date_end=".$l_line['date_end']."&p_exercice=".
 	    $l_line['p_exercice']."&$str_dossier\"> Changer</A>";
 	  $remove='<TD class="mtitle">';
-	  $remove.='<A class="mtitle" HREF="?p_action=periode&action=delete_per&p_per='.
-	    $l_line['p_id']."&$str_dossier\"> Efface</A>";
 
+
+	  if ($l_line['count_op'] == 0 ) {
+	    $remove.='<A class="mtitle" HREF="?p_action=periode&action=delete_per&p_per='.
+	      $l_line['p_id']."&$str_dossier\" onclick=\"return confirm('"._('Confirmez effacement ?')."')\" > Efface</A>";
+	  } else {
+	    $remove.=sprintf(_('Nombre opération %d'),$l_line['count_op']);
+	  }
+	  $remove.='</td>';
     }
 	echo "$closed";
 	echo $change;
